Change publishing so each channel.write gets its own PublishMessage
When there's more than one subscriber to a topic, either subscriber often
gets either a zero-length message or a message of the correct length that
only includes space characters instead of the correct body.

PublishMessage appears to be statefull - specifically, the ByteBuffer
payload.  Reusing the same ByteBuffer for different PublishMessages (or
reusing the same PublishMessage for different channels) sometimes leads to
the message body being either zero-length or "all spaces".  I assume this
is because ByteBuffer is statefull (stream-like).

Moquette/ProtocolProcessor creates a new PublishMessage for each message
sent to subscribers.